requirements.txt
torch>=1.10.0
torchvision>=0.11.0
streamlit>=1.0.0
streamlit-drawable-canvas>=0.8.0
Pillow>=9.0.0
numpy>=1.20.0
tqdm>=4.62.0